requirements.txt
absl-py
accelerate
adan-pytorch
aiohttp
aiosignal
aitextgen
antlr4-python3-runtime
asttokens
attrs
bash_kernel
beautifulsoup4
certifi
charset-normalizer
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
comm
contourpy
cycler
debugpy
decorator
diffusers==0.29.2
einops
executing
filelock
fire
fonttools
frozenlist
fsspec
ftfy
gdown
grpcio
huggingface-hub
hydra-core
idna
imageio
imhist
importlib_metadata
imWatermark
invisible-watermark
ipykernel
ipython
jedi
Jinja2
jupyter_client
jupyter_core
kiwisolver
lazy_loader
lightning-utilities
lmdb
lpips
Markdown
markdown-it-py
MarkupSafe
matplotlib
matplotlib-inline
mdurl
mpmath
multidict
nest-asyncio
networkx
numpy
omegaconf
open-clip-torch
opencv-python
packaging
pandas
parso
pexpect
pillow
platformdirs
prompt-toolkit
protobuf
psutil
ptyprocess
pure-eval
pyDeprecate
Pygments
pyparsing
PySocks
python-dateutil
pytorch-fid
pytorch-lightning
pytz
PyWavelets
PyYAML
pyzmq
regex
requests
rich
safetensors
scikit-image
scipy
sentencepiece
setuptools
six
soupsieve
stack-data
sympy
tensorboard
tensorboard-data-server
tensorboardX
termcolor
tifffile
timm
tokenizers
torch
torch-ema
torch-fidelity
pytorch-fid
pytorch-fid-wrapper
torchmetrics
torchvision
tornado
tqdm
traitlets
transformers
trimesh
typing_extensions
tzdata
urllib3
wandb
wcwidth
Werkzeug
wheel
yarl
zipp
